// app/types.d.ts
export type Movie = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string; 
    original_title: string;
    original_name: string;
    overview: string;
    popularity: Float;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: Float;
    vote_count: number;
}